+++ lib/flutter_eml_msg_viewer.dart
@@ -5,4 +5,4 @@ export 'src/model/mail_address.dart';
 export 'src/model/mail_attachment.dart';
 export 'src/model/mail_message.dart';
 export 'src/parser/mail_message_parser.dart';
-export 'src/widget/mail_message_viewer_widget.dart';
+export 'src/widget/mail_viewer_widget.dart';